Make LINUX_MUSL
